+++ src/Faker/Provider/HtmlLorem.php
@@ -2,6 +2,9 @@
 
 namespace Faker\Provider;
 
+use Faker\Provider\HtmlLorem\HtmlNode;
+use Faker\Provider\HtmlLorem\TextNode;
+
 class HtmlLorem extends Base
 {
 
@@ -19,11 +22,12 @@ class HtmlLorem extends Base
         $html->addNode($body);
         $head->addNode(HtmlLorem::randomTitle());
         $body->addNode(HtmlLorem::loginForm());
-        $this->randomSubTree($body,8);
+        $this->randomSubTree($body, 8);
         return $html->compile();
     }
 
-    private function randomSubTree(HtmlNode $root, $depth){
+    private function randomSubTree(HtmlNode $root, $depth)
+    {
         if($depth <= 0){
             return $root;
         }
@@ -40,11 +44,12 @@ class HtmlLorem extends Base
                 $root->addNode($sibling);
             }
 
-        }
+        };
         return $root;
     }
 
-    private function randomLeaf(){
+    private function randomLeaf()
+    {
         $node = new TextNode();
         $rand = mt_rand(1,10);
         switch($rand){
@@ -82,23 +87,27 @@ class HtmlLorem extends Base
         return $node;
     }
 
-    private function addRandomAttribute(HtmlNode $node){
+    private function addRandomAttribute(HtmlNode $node)
+    {
         $rand = mt_rand(1,100);
         if($rand <= 50){
             $node->addAttribute("class", Lorem::word());
         }
     }
 
-    public static function randomP(){
+    public static function randomP()
+    {
         return HtmlNode::newInstance("p")
             ->addNode(TextNode::newInstance(Lorem::sentence(mt_rand(1,30))));
     }
 
-    public static function randomText(){
+    public static function randomText()
+    {
         return TextNode::newInstance(Lorem::sentence(mt_rand(1,30)));
     }
 
-    public static function randomA(){
+    public static function randomA()
+    {
         $text = TextNode::newInstance()
             ->setText(Lorem::word());
 
@@ -108,57 +117,65 @@ class HtmlLorem extends Base
             ->addNode($text);
     }
 
-    public function randomTitle(){
+    public function randomTitle()
+    {
         $text = TextNode::newInstance()
-            ->setText(Lorem::sentence(mt_rand(1,3)));
+            ->setText(Lorem::sentence(mt_rand(1, 3)));
         return HtmlNode::newInstance()
             ->setType(HtmlNode::TITLE)
             ->addNode($text);
     }
 
-    public static function randomH1(){
+    public static function randomH1()
+    {
         $text = TextNode::newInstance()
-            ->setText(Lorem::sentence(mt_rand(1,10)));
+            ->setText(Lorem::sentence(mt_rand(1, 10)));
         return HtmlNode::newInstance(HtmlNode::H1)
             ->addNode($text);
     }
 
-    public static function randomH2(){
+    public static function randomH2()
+    {
         $text = TextNode::newInstance()
-            ->setText(Lorem::sentence(mt_rand(1,10)));
+            ->setText(Lorem::sentence(mt_rand(1, 10)));
         return HtmlNode::newInstance(HtmlNode::H2)
             ->addNode($text);
     }
 
-    public static function randomH3(){
+    public static function randomH3()
+    {
         $text = TextNode::newInstance()
-            ->setText(Lorem::sentence(mt_rand(1,10)));
+            ->setText(Lorem::sentence(mt_rand(1, 10)));
         return HtmlNode::newInstance(HtmlNode::H3)
             ->addNode($text);
     }
 
-    public static function randomB(){
+    public static function randomB()
+    {
         $text = TextNode::newInstance()
-            ->setText(Lorem::sentence(mt_rand(1,10)));
+            ->setText(Lorem::sentence(mt_rand(1, 10)));
         return HtmlNode::newInstance(HtmlNode::B)
             ->addNode($text);
     }
 
-    public static function randomI(){
+    public static function randomI()
+    {
         $text = TextNode::newInstance()
-            ->setText(Lorem::sentence(mt_rand(1,10)));
+            ->setText(Lorem::sentence(mt_rand(1, 10)));
         return HtmlNode::newInstance(HtmlNode::I)
             ->addNode($text);
     }
 
-    public static function randomSpan(){
+    public static function randomSpan()
+    {
         $text = TextNode::newInstance()
-            ->setText(Lorem::sentence(mt_rand(1,30)));
+            ->setText(Lorem::sentence(mt_rand(1, 30)));
         return HtmlNode::newInstance(HtmlNode::SPAN)
             ->addNode($text);
     }
 
-    public static function loginForm(){
+    public static function loginForm()
+    {
         $textInput = HtmlNode::newInstance(HtmlNode::INPUT)
             ->addAttribute("type", "text")
             ->addAttribute("id", "username");
@@ -184,7 +201,8 @@ class HtmlLorem extends Base
             ->addNode($submit);
     }
 
-    public static function randomTable(){
+    public static function randomTable()
+    {
         $cols = mt_rand(1,6);
         $rows = mt_rand(1,10);
 
@@ -202,7 +220,7 @@ class HtmlLorem extends Base
             $th  = HtmlNode::newInstance("th");
             $th->addNode(TextNode::newInstance(Lorem::sentence(mt_rand(1,4))));
             $tr->addNode($th);
-        }
+        };
         for($i = 0; $i < $rows; $i++){
             $tr = HtmlNode::newInstance("tr");
             $tbody->addNode($tr);
@@ -210,12 +228,13 @@ class HtmlLorem extends Base
                 $th  = HtmlNode::newInstance("td");
                 $th->addNode(TextNode::newInstance(Lorem::sentence(mt_rand(1,10))));
                 $tr->addNode($th);
-            }
-        }
+            };
+        };
         return $table;
     }
 
-    public static function randomUL(){
+    public static function randomUL()
+    {
         $num = mt_rand(1,10);
         $ul = HtmlNode::newInstance(HtmlNode::UL);
         for($i = 0; $i < $num; $i++){
@@ -226,134 +245,4 @@ class HtmlLorem extends Base
         return $ul;
     }
 
-}
-
-class HtmlNode implements Node
-{
-
-    const HTML = "html";
-    const HEAD = "head";
-    const BODY = "body";
-    const DIV = "div";
-    const A = "a";
-    const SPAN = "span";
-    const TABLE = "table";
-    const THEAD = "thead";
-    const TBODY = "tbody";
-    const UL = "ul";
-    const LI = "li";
-    const H1 = "h1";
-    const H2 = "h2";
-    const H3 = "h3";
-    const B= "b";
-    const I = "i";
-    const TITLE = "title";
-    const FORM = "form";
-    const INPUT = "input";
-    const LABEL = "label";
-
-    private $attributes = array();
-    /** @var Node[]  */
-    private $nodes = array();
-    private $type;
-
-    /**
-     * Document constructor.
-     * @param $type
-     */
-    public function __construct($type = null)
-    {
-        $this->type = $type;
-    }
-
-    /**
-     * @param null $type
-     */
-    public function setType($type)
-    {
-        $this->type = $type;
-        return $this;
-    }
-
-    /** @return HtmlNode */
-    public function addAttribute($key, $value){
-        $this->attributes[$key] = $value;
-        return $this;
-    }
-
-    /** @return HtmlNode */
-    public function addNode(Node $node)
-    {
-        $this->nodes[] = $node;
-        return $this;
-    }
-
-    public function compile()
-    {
-        if(!$this->type){
-            return "";
-        }
-        if(count($this->nodes) == 0){
-            return "<" . $this->type . $this->compileAttributes() . "/>";
-        }
-        return "<" . $this->type . $this->compileAttributes() . ">" . $this->compileNodes() .  "</" . $this->type . ">";
-    }
-
-    private function compileNodes()
-    {
-        $nodes = "";
-        foreach($this->nodes as $node){
-            $nodes = $nodes . $node->compile();
-        }
-        return $nodes;
-    }
-
-    private function compileAttributes()
-    {
-        $attributes = "";
-        foreach($this->attributes as $key => $value){
-            $attributes = $attributes . " " . $key . "=\"" . $value . "\"";
-        }
-        return $attributes;
-    }
-
-    /** @return HtmlNode */
-    public static function newInstance($type = null){
-        $node = new HtmlNode($type);
-        return $node;
-    }
-
-}
-
-class TextNode implements Node{
-
-    private $text = "";
-
-    public function __construct($text = null)
-    {
-        $this->text = $text;
-    }
-
-    function compile()
-    {
-        return $this->text;
-    }
-
-    /** @return TextNode */
-    public function setText($text)
-    {
-        $this->text = $text;
-        return $this;
-    }
-
-    /** @return TextNode */
-    public static function newInstance($text = null){
-        return new TextNode($text);
-    }
-
-}
-
-interface Node
-{
-    function compile();
 }
