+++ src/Faker/Provider/HtmlLorem/Node.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Created by IntelliJ IDEA.
+ * User: Play
+ * Date: 7/20/2016
+ * Time: 8:38 PM
+ */
+
+namespace Faker\Provider\HtmlLorem;
+
+interface Node
+{
+    public function compile();
+}
